+++ src/main/java/store/demostore/controllers/SaleController.java
@@ -2,11 +2,14 @@ package store.demostore.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import jakarta.validation.Valid;
 import store.demostore.models.dto.SaleDto;
 import store.demostore.models.entities.SaleEntity;
 import store.demostore.services.SaleService;
+import store.demostore.services.handleErrors.HandleErrorsService;
 import store.demostore.utils.Constants;
 
 import java.util.List;
@@ -18,27 +21,29 @@ public class SaleController {
     @Autowired
     private SaleService saleService;
 
+    @Autowired
+    private HandleErrorsService habErrorsService;
+
     @GetMapping
     public List<SaleDto> findAll() {
         return saleService.findAll();
     }
 
-    /*
-     * 
-     * @GetMapping("/{id}")
-     * public ResponseEntity<?> findById(@PathVariable String id) {
-     * return saleService.findById(id);
-     * }
-     * 
-     */
-
     @PostMapping
-    public ResponseEntity<?> save(@RequestBody SaleEntity sale) {
+    public ResponseEntity<?> save(@Valid @RequestBody SaleEntity sale, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return habErrorsService.vaidateErrors(bindingResult);
+
+        }
         return saleService.save(sale);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> delete(@PathVariable String id) {
+    public ResponseEntity<?> delete(@Valid @PathVariable String id, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return habErrorsService.vaidateErrors(bindingResult);
+
+        }
         return saleService.remove(id);
     }
